python_app/requirements.txt
numpy
nibabel
pillow
opencv-python-headless
